Refactored attendance component a bit
