+++ packages/attendance/src/components/simple/TimeBar/Display.tsx
@@ -23,29 +23,32 @@ export const Display: React.FC<any> = ({
     <Box bg="white" p="1" {..._box}>
       <HStack justifyContent="space-between" alignItems="center" space={4}>
         <HStack space="4" alignItems="center">
-          <IconByName
-            size="50"
-            color={
-              typeof previousDisabled === "undefined" ||
-              previousDisabled === false
-                ? activeColor
-                  ? activeColor
-                  : colors.primary
-                : // @ts-ignore
-                  colors.grayInLight
-            }
-            name="ArrowLeftSLineIcon"
-            onPress={(e) => {
-              if (leftErrorText) {
-                toast.show(leftErrorText);
-              } else if (
+          {
+            // @ts-ignore
+            <IconByName
+              size="50"
+              color={
                 typeof previousDisabled === "undefined" ||
                 previousDisabled === false
-              ) {
-                setPage(page - 1);
+                  ? activeColor
+                    ? activeColor
+                    : colors.primary
+                  : // @ts-ignore
+                    colors.grayInLight
               }
-            }}
-          />
+              name="ArrowLeftSLineIcon"
+              onPress={(e) => {
+                if (leftErrorText) {
+                  toast.show(leftErrorText);
+                } else if (
+                  typeof previousDisabled === "undefined" ||
+                  previousDisabled === false
+                ) {
+                  setPage(page - 1);
+                }
+              }}
+            />
+          }
         </HStack>
         <HStack space="4" alignItems="center">
           <Text fontSize="md" bold>
@@ -53,27 +56,30 @@ export const Display: React.FC<any> = ({
           </Text>
         </HStack>
         <HStack space="2">
-          <IconByName
-            size="50"
-            color={
-              typeof nextDisabled === "undefined" || nextDisabled === false
-                ? activeColor
+          {
+            // @ts-ignore
+            <IconByName
+              size="50"
+              color={
+                typeof nextDisabled === "undefined" || nextDisabled === false
                   ? activeColor
-                  : colorTheme.gray
-                : colorTheme.grayIndark
-            }
-            name="ArrowRightSLineIcon"
-            onPress={(e) => {
-              if (rightErrorText) {
-                toast.show(rightErrorText);
-              } else if (
-                typeof nextDisabled === "undefined" ||
-                nextDisabled === false
-              ) {
-                setPage(page + 1);
+                    ? activeColor
+                    : colorTheme.gray
+                  : colorTheme.grayIndark
               }
-            }}
-          />
+              name="ArrowRightSLineIcon"
+              onPress={(e) => {
+                if (rightErrorText) {
+                  toast.show(rightErrorText);
+                } else if (
+                  typeof nextDisabled === "undefined" ||
+                  nextDisabled === false
+                ) {
+                  setPage(page + 1);
+                }
+              }}
+            />
+          }
         </HStack>
       </HStack>
     </Box>
