+++ packages/attendance/src/pages/Attendance.js
@@ -18,7 +18,6 @@ import { Box, FlatList, HStack, Stack, VStack, Button } from "native-base";
 import { TimeBar } from "components/simple/TimeBar/TimeBar";
 import { GetAttendance } from "../services/calls/registryCalls";
 
-
 import { MultipleAttendance } from "components/composite/MultipleAttendance";
 import AttendanceComponent from "components/composite/AttendanceComponent";
 import moment from "moment";
