+++ packages/attendance/src/utils/functions/GetLastAttendance.ts
@@ -1,0 +1,6 @@
+import moment from "moment";
+export const GetLastAttendance = (attendance) => {
+  let dates = attendance.map((d) => moment(d.updatedAt));
+  let date = moment.max(dates);
+  return dates.length ? date.format("hh:mmA") : "N/A";
+};
