+++ packages/attendance/src/utils/functions/GetStatusFromManifest.ts
@@ -1,0 +1,11 @@
+export const GetStatusFromManifest = (manifest) => {
+  const status = Array.isArray(
+    manifest?.["attendance.default_attendance_states"]
+  )
+    ? manifest?.["attendance.default_attendance_states"]
+    : manifest?.["attendance.default_attendance_states"]
+    ? JSON.parse(manifest?.["attendance.default_attendance_states"])
+    : [];
+
+  return status;
+};
