+++ packages/attendance/tsconfig.json
@@ -2,12 +2,12 @@
   "include": ["src"],
   "exclude": ["node_modules", "*/**/*.json"],
   "compilerOptions": {
+    "baseUrl": "./src",
     "noImplicitAny": false,
     "jsx": "react",
     "allowJs": true,
     "resolveJsonModule": true,
     "esModuleInterop": true,
-    "skipLibCheck": true,
-    "baseUrl": "src",
+    "skipLibCheck": true
   }
 }
